only have one config template and use mustache to adjust them
include writing correct usernames, fix #130
