+++ pkg.vars.config
@@ -35,3 +35,11 @@
  "specific language governing permissions and limitations\n"
  "under the License."}.
 {solaris_pkgname, "INDIGOtts"}.
+%%
+%%
+{idh_script, "idh/basic-idh.py"}.
+{oidc_info, "#please add your client id and secret and rename this file to .conf"}.
+{oidc_google_id, "<replace with your client id>"}.
+{oidc_google_secret, "<replace with your client secret>"}.
+{oidc_iam_id, "<replace with your client id>"}.
+{oidc_iam_secret, "<replace with your client secret>"}.
