+++ rebar.config
@@ -26,11 +26,17 @@
         {include_src, false},
         {generate_start_script, false},
         {overlay, [
-                   {mkdir, "etc/oidc"},
-                   {mkdir, "etc/services"},
-                   {template, "tts_config/rel/tts/main.conf","etc/main.conf"},
+                   {mkdir, "etc"},
                    {copy, "config/vm.args","etc/vm.args"},
                    {template, "config/sys.config","etc/app.config"},
+
+                   {mkdir, "etc/oidc"},
+                   {template, "tts_config/oidc/google.conf","etc/oidc/google.sample"},
+                   {template, "tts_config/oidc/iam.conf","etc/oidc/iam.sample"},
+                   {mkdir, "etc/services"},
+                   {template, "tts_config/services/ssh.conf","etc/services/ssh.sample"},
+                   {template, "tts_config/services/opennebula.conf","etc/services/opennebula.sample"},
+                   {template, "tts_config/main.conf","etc/main.conf"},
 
                    {mkdir, "bin"},
                    {copy, "client/ttsc.sh","bin/ttsc"},
@@ -40,15 +46,14 @@
                    {template, "_build/default/lib/node_package/priv/base/nodetool", "erts-{{erts_vsn}}/bin/nodetool"},
                    {template, "_build/default/lib/node_package/priv/base/env.sh", "lib/env.sh"},
 
-                   {mkdir, "share/tmp"},
-
                    {mkdir, "data/plugins"},
-                   {copy, "tts_config/rel/tts/scripts/ssh.py", "data/plugins/ssh.py"},
-                   {copy, "tts_config/rel/tts/scripts/opennebula.py", "data/plugins/opennebula.py"},
+                   {template, "tts_config/plugins/ssh.py", "data/plugins/ssh.py"},
+                   {template, "tts_config/plugins/x509.py", "data/plugins/x509.py"},
+                   {template, "tts_config/plugins/opennebula.py", "data/plugins/opennebula.py"},
 
                    {mkdir, "data/idh"},
-                   {template, "tts_config/rel/tts/scripts/basic-idh.py", "data/idh/basic-idh.py"},
-                   {copy, "tts_config/rel/tts/scripts/create_user.py", "data/idh/create_user.py"}
+                   {template, "tts_config/scripts/basic-idh.py", "data/idh/basic-idh.py"},
+                   {template, "tts_config/scripts/create_user.py", "data/idh/create_user.py"}
                   ]}
        ]}.
