--- tts_config/README.md
@@ -1,4 +1,0 @@
-# Sample and local development configuration
-Copy the sample/tts folder to `~/.config/tts`
-build and start TTS and you can create ssh keys on your local machine (if you
-have ssh installed)
