+++ tts_config/main.conf
@@ -1,5 +1,5 @@
 HostName =  localhost 
 SSL = false
 Port = 8080 
-IDHScript = /usr/share/tts/idh/basic-idh.py
+IDHScript = {{platform_data_dir}}/{{idh_script}}
 SqliteFile = {{platform_data_dir}}/tts.db
