+++ tts_config/oidc/google.conf
@@ -1,0 +1,7 @@
+{{oidc_info}}
+
+Id = Google 
+Description = Google, the well known search enging 
+ClientId =  {{oidc_google_id}} 
+Secret =  {{oidc_google_secret}} 
+ConfigEndpoint = https://accounts.google.com/.well-known/openid-configuration
