+++ tts_config/oidc/iam.conf
@@ -1,5 +1,7 @@
+{{oidc_info}}
+
 Id = IAM
 Description = INDIGO Datacloud Identity and Access Management (IAM)
-ClientId =  e70e7190-f19a-44be-8f6c-d64cad3771b9
-Secret =  AN5pY379_Z_1jgiiAzmQFB4lYMsBYJOAQrCoTtf2stPA6dMghW5KAoKob4zCDCg_yKBEHe8zKP2KuiWxV34zWzw
+ClientId =  {{oidc_iam_id}} 
+Secret =  {{oidc_iam_secret}} 
 ConfigEndpoint = https://iam-test.indigo-datacloud.eu/.well-known/openid-configuration
