--- tts_config/sample/tts/main.conf
@@ -1,4 +1,0 @@
-HostName =  localhost 
-SSL = false
-Port = 8080 
-IDHScript = ./scripts/idh-single-user.py
