--- tts_config/sample/tts/oidc/google.conf
@@ -1,5 +1,0 @@
-Id = Google 
-Description = Google, the well known search enging ... and more 
-ClientId =  65375832888-m99kcr0vu8qq95h588b1rhi52ei234qo.apps.googleusercontent.com 
-Secret =  MEfMXcaQtckJPBctTrAuSQkJ 
-ConfigEndpoint = https://accounts.google.com/.well-known/openid-configuration
