+++ tts_config/services/ssh.conf
@@ -4,7 +4,7 @@ Host = localhost
 Port = 22
 Description = local ssh access to your own machine 
 
-Cmd = ~/.config/tts/scripts/ssh.py 
+Cmd = {{platform_data_dir}}/plugins/ssh.py 
 
 ConnectionType = local 
 CredentialLimit = 3
