+++ utils/install_sample_config
@@ -1,11 +1,43 @@
 #!/bin/bash
-DIRECTORY=~/.config/tts
+DATA_DIR=~/.config/tts
 BACKUPDIR=~/.config/tts_`date +%Y%m%d_%H%M%S`
 
-if [ -d "$DIRECTORY" ]; then
+if [ -d "$DATA_DIR" ]; then
     echo "storing backup of current config in $BACKUPDIR"
-    mv "$DIRECTORY" "$BACKUPDIR"
+    mv "$DATA_DIR" "$BACKUPDIR"
 fi
 
-echo "copy configuration"
-cp -vr ./tts_config/sample/tts ~/.config/tts
+echo -n "copy configuration "
+export platform_data_dir=$DATA_DIR
+export idh_script="scripts/idh-single-user.py"
+export oidc_info="\
+# WARNING: these settings only allow http://localhost:8080\
+# NOT MENT FOR PRODUCTION!"
+export oidc_google_id="65375832888-m99kcr0vu8qq95h588b1rhi52ei234qo.apps.googleusercontent.com"
+export oidc_google_secret="MEfMXcaQtckJPBctTrAuSQkJ"
+export oidc_iam_id="e70e7190-f19a-44be-8f6c-d64cad3771b9"
+export oidc_iam_secret="AN5pY379_Z_1jgiiAzmQFB4lYMsBYJOAQrCoTtf2stPA6dMghW5KAoKob4zCDCg_yKBEHe8zKP2KuiWxV34zWzw"
+DIRS="./
+oidc/
+scripts/
+plugins/
+services/"
+
+for d in $DIRS 
+do
+    OutDir=`realpath -m $DATA_DIR/$d`
+    mkdir -p $OutDir
+    # echo "*** $OutDir"
+    for f in ./tts_config/$d*
+    do
+        echo -n "."
+        FileName=`basename $f`
+        InFile=`realpath $f`
+        OutFile=$OutDir/$FileName
+        if [ -f $InFile ]; then
+            # echo "   $InFile -> $OutFile"
+            cat $InFile | ./utils/mo > $OutFile
+        fi
+    done
+done
+echo " done"
